Add new test case and update guidance text
Added a new test case to `tests_list.txt` for the conversion of the number 2024 to its Roman numeral representation. Updated the `gpt_guidance.txt` with clearer instructions on how tests and implementation code should be presented for incorporation into their respective files. Also added guidance on how changes to the code should be highlighted and reviewed once a list of tests is complete.

Signed-off-by: Sean Moir <sean.moir@facilligent.com>